refactor: Regroup doctypes
