--- src/doctypes/bank.js
@@ -1,5 +1,0 @@
-const BANK_DOCTYPE = 'io.cozy.bank.operations'
-
-module.exports = {
-  BANK_DOCTYPE
-}
