--- src/doctypes/jobs.js
@@ -1,5 +1,0 @@
-const JOBS_DOCTYPE = 'io.cozy.jobs'
-
-module.exports = {
-  JOBS_DOCTYPE
-}
