--- src/doctypes/models.js
@@ -1,1 +1,0 @@
-export const MODELS_DOCTYPE = 'io.cozy.dissec.models'
