--- src/doctypes/nodes.js
@@ -1,5 +1,0 @@
-const NODES_DOCTYPE = 'io.cozy.dissec.nodes'
-
-module.exports = {
-  NODES_DOCTYPE
-}
