--- src/doctypes/observations.js
@@ -1,5 +1,0 @@
-const OBSERVATIONS_DOCTYPE = 'io.cozy.dissec.observations'
-
-module.exports = {
-  OBSERVATIONS_DOCTYPE
-}
