--- src/doctypes/shares.js
@@ -1,1 +1,0 @@
-export const SHARES_DOCTYPE = 'io.cozy.dissec.shares'
