--- src/doctypes/triggers.js
@@ -1,5 +1,0 @@
-const TRIGGERS_DOCTYPE = 'io.cozy.triggers'
-
-module.exports = {
-  TRIGGERS_DOCTYPE
-}
